+++ CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 - Added metrics for periodicRegisterer [#18](https://github.com/xmidt-org/wrp-listener/pull/18)
+- Added parser for extracting device ID from wrp message [#23](https://github.com/xmidt-org/wrp-listener/pull/23)
 
 ## [v0.1.2]
 - Added travis automation for github releases [#13](https://github.com/xmidt-org/wrp-listener/pull/13)